+++ packages/define-plugin/src/utils.ts
@@ -1,5 +1,4 @@
 import path from 'node:path'
-import shell from 'shelljs'
 import minimist from 'minimist'
 import { isNonEmptyString } from '@txjs/bool'
 
@@ -8,12 +7,14 @@ const ciArgs = minimist(process.argv.slice(2), {
   boolean: ['watch'],
 })
 
+export const platforms = ['weapp', 'alipay', 'tt'] as const
+
 export type NodeEnv =
   | 'development'
   | 'production'
   | (string & Record<never, never>)
 
-export type MipraEnv = 'weapp' | 'alipay' | 'tt'
+export type MipraEnv = (typeof platforms)[number]
 
 export const isWatch = () => {
   return !!ciArgs.watch
@@ -39,29 +40,13 @@ export const processResolve = (...dir: string[]) => {
   return resolve(process.cwd(), ...dir)
 }
 
-export const preprocessPath = (outputPath: string, path: string) => {
-  let tempPath = path
-  const foundAt = path.lastIndexOf('/')
-
-  if (foundAt > 0) {
-    const folder = path.slice(0, foundAt + 1)
-    const folderPath = processResolve(outputPath, folder)
-    if (!shell.test('-d', folderPath)) {
-      shell.mkdir('-p', folderPath)
-    }
-    tempPath = path.slice(0, foundAt)
-  }
-  return tempPath
-}
-
 export const toJson = (inputString?: string) => {
   try {
     if (inputString) {
       return JSON.parse(inputString)
     }
     throw new Error('inputString is empty.')
-  } catch (error) {
-    console.error(error)
+  } catch {
     return null
   }
 }