+++ packages/env-plugin/src/env.ts
@@ -1,8 +1,9 @@
+import { platforms } from '@mipra-helper/define-plugin'
 import {
   DeclareTypes,
   type DeclareTypesOption,
 } from '@mipra-helper/declare-plugin'
-import { json2ts } from 'json-ts'
+import { json2tsMulti } from 'json-ts'
 import { dotenv } from './dotenv'
 
 export class EnvTypes extends DeclareTypes {
@@ -17,10 +18,16 @@ export class EnvTypes extends DeclareTypes {
   override eventName = 'onEnvUpdate'
 
   override build(): string {
-    const content = json2ts(JSON.stringify(dotenv.loadEnv()), {
-      prefix: '',
-      rootName: 'ProcessEnv',
-    })
-    return `namespace NodeJS { ${content} }`
+    const envInferface = json2tsMulti(
+      [JSON.stringify({ TARO_ENV: '' }), JSON.stringify(dotenv.loadEnv())],
+      {
+        prefix: '',
+        rootName: 'ProcessEnv',
+      }
+    ).replace(
+      /(\n\s+TARO_ENV\?:\s+)(\w+);\n/,
+      `$1${platforms.map((platform) => `'${platform}'`).join(' | ')};\n`
+    )
+    return `namespace NodeJS { ${envInferface} }`
   }
 }