+++ packages/runner-plugin/src/index.ts
@@ -9,10 +9,15 @@ export default definePlugin<RunnerPluginOption>(
     option.outputPlatform ||= true
 
     ctx.modifyRunnerOpts(({ opts }) => {
-      if (option.outputPlatform) {
+      if (
+        option.outputPlatform &&
+        !opts.outputRoot.endsWith(ctx.mipraEnv) &&
+        !opts.outputRoot.endsWith(ctx.nodeEnv)
+      ) {
         opts.outputRoot = `${opts.outputRoot}/${ctx.mipraEnv}`
       }
       ctx.logger('Configuration modification completed')
+      return opts
     })
   },
   {