+++ packages/condition-plugin/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mipra-helper/condition-plugin
 
+## 0.0.5
+
+### Patch Changes
+
+- fix 修复部分问题
+- Updated dependencies
+  - @mipra-helper/define-plugin@0.0.4
+  - @mipra-helper/env-plugin@0.0.4
+  - @mipra-helper/tsconfig@0.0.3
+
 ## 0.0.4
 
 ### Patch Changes