+++ packages/define-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mipra-helper/define-plugin
 
+## 0.0.4
+
+### Patch Changes
+
+- fix 修复部分问题
+- Updated dependencies
+  - @mipra-helper/tsconfig@0.0.3
+
 ## 0.0.3
 
 ### Patch Changes