+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mipra-helper/tsconfig
 
+## 0.0.3
+
+### Patch Changes
+
+- fix 修复部分问题
+
 ## 0.0.2
 
 ### Patch Changes